+++ src/common/index.ts
@@ -1,9 +1,6 @@
-import { AnchorProvider, web3, Program, utils } from '@project-serum/anchor';
-import { Spl, SPL_ACCOUNT_LAYOUT } from '@raydium-io/raydium-sdk';
+import { web3, utils } from '@project-serum/anchor';
 
 import { NodeWallet } from './classes/nodewallet';
-import { AccountInfoData, AccountInfoParsed, GetTokenAccount, ParseTokenAccount, UserNFT } from './types';
-import idl from './idl/multi_reward_staking.json';
 
 //when we only want to view vaults, no need to connect a real wallet.
 export const createFakeWallet = () => {
@@ -85,62 +82,4 @@ export const createAssociatedTokenAccountInstruction = (
       data: Buffer.from([]),
     }),
   ];
-};
-
-export const deriveMetadataPubkeyFromMint = async (nftMint: web3.PublicKey): Promise<web3.PublicKey> => {
-  let metadata_program = new web3.PublicKey('metaqbxxUerdq28cj1RbAWkYQm3ybzjb6a8bt518x1s');
-
-  const encoder = new TextEncoder();
-  const [metadataPubkey] = await web3.PublicKey.findProgramAddress(
-    [encoder.encode('metadata'), metadata_program.toBuffer(), nftMint.toBuffer()],
-    metadata_program,
-  );
-
-  return metadataPubkey;
-};
-
-export const decodeSplTokenAccountData = (tokenAccountDataEncoded: Buffer): AccountInfoData =>
-  SPL_ACCOUNT_LAYOUT.decode(tokenAccountDataEncoded);
-
-export const parseTokenAccount: ParseTokenAccount = ({ tokenAccountPubkey, tokenAccountEncoded }) =>
-  tokenAccountEncoded
-    ? {
-        pubkey: tokenAccountPubkey,
-        accountInfo: decodeSplTokenAccountData(tokenAccountEncoded.data),
-      }
-    : null;
-
-export const getTokenAccount = async ({
-  tokenMint,
-  owner,
-  connection,
-}: GetTokenAccount): Promise<AccountInfoParsed | null> => {
-  const tokenAccountPubkey = await Spl.getAssociatedTokenAccount({
-    mint: tokenMint,
-    owner,
-  });
-
-  const tokenAccountEncoded = await connection.getAccountInfo(tokenAccountPubkey);
-
-  return parseTokenAccount({ tokenAccountPubkey, tokenAccountEncoded });
-};
-
-export const getTokenAccountBalance = (lpTokenAccountInfo: AccountInfoParsed, lpDecimals: number): number =>
-  lpTokenAccountInfo?.accountInfo?.amount.toNumber() / 10 ** lpDecimals || 0;
-
-export const shortenAddress = (address: string, chars = 4): string =>
-  `${address.slice(0, chars)}...${address.slice(-chars)}`;
-
-export const getNftCreators = (nft: UserNFT): string[] =>
-  nft?.metadata?.properties?.creators?.filter(({ verified }) => verified)?.map(({ address }) => address) || [];
-
-export const returnAnchorMultiRewardStaking = async (
-  programId: web3.PublicKey,
-  connection: web3.Connection,
-): Promise<Program> => {
-  return new Program(
-    idl as any,
-    programId,
-    new AnchorProvider(connection, createFakeWallet(), AnchorProvider.defaultOptions()),
-  );
 };
