+++ src/index.ts
@@ -1,13 +1,10 @@
 export * from './common/types';
 export * from './loans/types';
 export * from './pools/types';
-export * from './swaps/types';
 
 export * as utils from './common';
 export * as loans from './loans';
 export * as pools from './pools';
-export * as swaps from './swaps';
+
 export * as raydium from '@raydium-io/raydium-sdk';
-export { AnchorProvider } from '@project-serum/anchor';
-export { BN } from '@project-serum/anchor';
-export { web3 } from '@project-serum/anchor';
+export { AnchorProvider, BN, web3 } from '@project-serum/anchor';
