+++ src/pools/constants.ts
@@ -1,3 +1,3 @@
-import { web3 } from '@project-serum/anchor';
-
 export const ACCOUNT_PREFIX = 'communitypools';
+
+export const BLOCKED_POOLS_IDS = ['DqMPkcRT22dPTTsj2rxhCNxg44D93NXiSBWiACAAXd2Q'];
