+++ src/pools/index.ts
@@ -34,3 +34,4 @@ export * from './staking/harvestSecondaryReward';
 export * from './staking/stakeInFusion';
 export * from './staking/getAllProgramStakingAccounts';
 
+export * from './helpers';
