+++ src/pools/read_functions/getWhitelistedCreatorsDictionary.ts
@@ -1,6 +1,6 @@
 import { Dictionary } from 'lodash';
 
-import { PoolWhitelistType } from '../../common/types';
+import { PoolWhitelistType } from '../types';
 
 export const getWhitelistedCreatorsDictionary = (pool): Dictionary<boolean> =>
   Object.fromEntries(
