+++ src/pools/read_functions/isNFTWhitelistedByMint.ts
@@ -1,7 +1,11 @@
 import { Dictionary } from 'lodash';
-import { UserNFT } from '../../common/types';
 
-export const isNFTWhitelistedByMint = (nft: UserNFT, whitelistedMintsDictionary: Dictionary<boolean>): string | null => {
+import { UserNFT } from '../types';
+
+export const isNFTWhitelistedByMint = (
+  nft: UserNFT,
+  whitelistedMintsDictionary: Dictionary<boolean>,
+): string | null => {
   const { mint } = nft;
 
   return whitelistedMintsDictionary[mint] ? mint : null;
