+++ src/pools/staking/unstakeInFusion.ts
@@ -1,6 +1,7 @@
-import { AnchorProvider, BN, web3, utils } from '@project-serum/anchor';
+import { BN, web3, utils } from '@project-serum/anchor';
 
-import { findAssociatedTokenAddress, returnAnchorMultiRewardStaking } from '../../common';
+import { findAssociatedTokenAddress } from '../../common';
+import { returnAnchorMultiRewardStaking } from '../helpers';
 
 export const unstakeInFusion = async (
   programId: web3.PublicKey,
