+++ src/pools/types.ts
@@ -1,6 +1,27 @@
-import { BN, AnchorProvider, web3 } from '@project-serum/anchor';
+import { BN, web3 } from '@project-serum/anchor';
+import { LiquidityPoolInfo, LiquidityPoolKeysV4, Percent } from '@raydium-io/raydium-sdk';
 
-import { PoolDataByMint, TokenInfo } from '../common/types';
+import { TokenInfo } from '../common/types';
+
+export type RaydiumPoolInfoMap = Map<string, LiquidityPoolInfo>;
+
+export interface GetInputAmount {
+  poolKeys: LiquidityPoolKeysV4;
+  poolInfo: LiquidityPoolInfo;
+  receiveToken: TokenInfo;
+  receiveAmount: number;
+  payToken: TokenInfo;
+  slippage?: Percent;
+}
+
+export interface GetOutputAmount {
+  poolKeys: LiquidityPoolKeysV4;
+  poolInfo: LiquidityPoolInfo;
+  payToken: TokenInfo;
+  payAmount: number;
+  receiveToken: TokenInfo;
+  slippage?: Percent;
+}
 
 export interface BoardEntryView {
   boardEntryPubkey: string;
@@ -400,4 +421,83 @@ export interface SecondaryRewardView {
   decimalsOutput: string;
   startTime: string;
   endTime: string;
+}
+
+export interface AccountInfoData {
+  owner: web3.PublicKey;
+  state: number;
+  mint: web3.PublicKey;
+  amount: BN;
+  delegateOption: number;
+  delegate: web3.PublicKey;
+  isNativeOption: number;
+  isNative: BN;
+  delegatedAmount: BN;
+  closeAuthorityOption: number;
+  closeAuthority: web3.PublicKey;
+}
+
+export interface AccountInfoParsed {
+  pubkey: web3.PublicKey;
+  accountInfo: AccountInfoData;
+}
+
+export type ParseTokenAccount = (params: {
+  tokenAccountPubkey: web3.PublicKey;
+  tokenAccountEncoded: web3.AccountInfo<Buffer> | null;
+}) => AccountInfoParsed | null;
+
+export interface GetTokenAccount {
+  tokenMint: web3.PublicKey;
+  owner: web3.PublicKey;
+  connection: web3.Connection;
+}
+
+export interface NFTCreator {
+  address: string;
+  share: number;
+  verified?: boolean;
+}
+
+interface NFTFile {
+  type: string;
+  uri: string;
+}
+
+export interface ArweaveAttribute {
+  trait_type: string;
+  value: number | string;
+}
+
+export interface ArweaveMetadata {
+  name: string;
+  symbol: string;
+  description: string;
+  collectionName?: string;
+  seller_fee_basis_points?: number;
+  image: string;
+  animation_url?: string;
+  external_url?: string;
+  attributes: ArweaveAttribute[];
+  properties?: {
+    creators?: NFTCreator[];
+    files?: NFTFile[];
+  };
+}
+
+export interface UserNFT {
+  mint: string;
+  metadata: ArweaveMetadata;
+}
+
+export interface PoolData {
+  tokenInfo: TokenInfo;
+  poolConfig: LiquidityPoolKeysV4;
+}
+
+export type PoolDataByMint = Map<string, PoolData>;
+
+export enum PoolWhitelistType {
+  SINGLE_NFT_WHITELIST = 'singleNftWhitelist',
+  CREATOR_WHITELIST = 'creatorWhitelist',
 }
