--- src/swaps/index.ts
@@ -1,4 +1,0 @@
-export * from './read_functions/fetchRaydiumPoolsInfo';
-export * from './read_functions/fetchRaydiumPoolsInfoMap';
-export * from './read_functions/getInputAmount';
-export * from './read_functions/getOutputAmount';
